+++ app/src/main/java/com/github/damontecres/wholphin/ui/playback/PlaybackViewModel.kt
@@ -864,53 +864,3 @@ private fun applyTrackSelections(
         }
     }
 }
-
-private fun determineSubtitleIndex(
-    subtitleStreams: List<SubtitleStream>,
-    subtitleMode: SubtitlePlaybackMode,
-    audioLanguage: String?,
-    subtitleLanguage: String?,
-    playbackConfig: ItemPlayback,
-): Int? {
-    if (playbackConfig.subtitleIndex == TrackIndex.DISABLED) {
-        return null
-    } else if (playbackConfig.subtitleIndexEnabled) {
-        return playbackConfig.subtitleIndex
-    } else {
-        return when (subtitleMode) {
-            SubtitlePlaybackMode.ALWAYS -> {
-                if (subtitleLanguage != null) {
-                    subtitleStreams.firstOrNull { it.language == subtitleLanguage }?.index
-                } else {
-                    subtitleStreams.firstOrNull()?.index
-                }
-            }
-
-            SubtitlePlaybackMode.ONLY_FORCED ->
-                if (subtitleLanguage != null) {
-                    subtitleStreams.firstOrNull { it.language == subtitleLanguage && it.forced }?.index
-                } else {
-                    subtitleStreams.firstOrNull { it.forced }?.index
-                }
-
-            SubtitlePlaybackMode.SMART -> {
-                if (audioLanguage != null && subtitleLanguage != null && audioLanguage != subtitleLanguage) {
-                    subtitleStreams.firstOrNull { it.language == subtitleLanguage }?.index
-                } else {
-                    null
-                }
-            }
-
-            SubtitlePlaybackMode.DEFAULT -> {
-                // TODO check for language?
-                (
-                    subtitleStreams.firstOrNull { it.default && it.forced }
-                        ?: subtitleStreams.firstOrNull { it.default }
-                        ?: subtitleStreams.firstOrNull { it.forced }
-                )?.index
-            }
-
-            SubtitlePlaybackMode.NONE -> null
-        }
-    }
-}
